Update path since db/cache package moved
