+++ server/server.go
@@ -23,10 +23,10 @@ import (
 
 	"github.com/kardiachain/explorer-backend/types"
 
+	"github.com/kardiachain/explorer-backend/cache"
+	"github.com/kardiachain/explorer-backend/db"
 	"github.com/kardiachain/explorer-backend/kardia"
 	"github.com/kardiachain/explorer-backend/metrics"
-	"github.com/kardiachain/explorer-backend/server/cache"
-	"github.com/kardiachain/explorer-backend/server/db"
 )
 
 type Config struct {
